projeto/requirements.txt
appdirs==1.4.3
colorama==0.3.7
cycler==0.10.0
decorator==4.0.11
ipython==5.3.0
ipython-genutils==0.2.0
matplotlib==2.0.0
nltk==3.2.2
numpy==1.12.1
packaging==16.8
pickleshare==0.7.4
prompt-toolkit==1.0.13
pyenchant==1.6.8
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
